requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0

# Database
sqlalchemy>=1.4.0
pymysql>=1.0.0

# Configuration
python-dotenv>=0.19.0

# Data validation
pydantic>=1.10.0

# XML processing
lxml>=4.6.0

# Date/time handling
python-dateutil>=2.8.0

# Logging and utilities
colorlog>=6.6.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Type checking (optional but recommended)
mypy>=1.0.0
